Basic Baseball Stat Analysis
Grabbing the stats from the Lahman database and exploring relationships/visualisations for Wins as a function of runs, salary, and homers. plus other things like that. Had an idea to try and make a model to predict WAr and found materials, pybaseball needs to be updated since the imports I'd like to use are broken. will try a work around with the imports that do work